backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "api"
authors = ["shun198"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.24.2"
mysqlclient = "^2.1.1"
gunicorn = "^20.1.0"
drf-nested-routers = "^0.93.4"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
redis = "^4.3.4"
django-cors-headers = "^3.13.0"
Authlib = "^1.1.0"
django-filter = "^22.1"
boto3 = "^1.26.22"
injector = "^0.20.1"
django-ses = "^3.2.2"
django-storages = "^1.13.1"
pydantic = "^1.10.6"
pillow = "^9.4.0"
toml = "^0.10.2"
kombu = "^5.2.4"
pycurl = "^7.45.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.6"
pytest-xdist = "^3.0.2"
django-debug-toolbar = "^3.8.1"
pytest-bdd = "^6.1.1"
pytest-clarity = "^1.0.1"
pytest-custom-exit-code = "^0.3.0"
factory-boy = "^3.2.1"
pytest-mock = "^3.10.0"
black = "^22.10.0"
isort = "^5.11.4"
django-extensions = "^3.2.1"
pdoc = "^13.0.0"

[tool.isort]
profile = "black"

[tool.black]
line-length = 79
include = '\.py$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v -s --durations=0"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "project.settings.local"

# logの設定
[logging]
version = 1
[logging.formatters.simple]
format = "[%(levelname)s] %(name)s - %(message)s "
[logging.handlers.applicationHandler]
class = "logging.StreamHandler"
level = "INFO"
formatter = "simple"
stream = "ext://sys.stdout"
[logging.handlers.errorHandler]
class = "logging.StreamHandler"
level = "ERROR"
formatter = "simple"
stream = "ext://sys.stdout"
[logging.loggers.application]
level = "DEBUG"
handlers = ["applicationHandler"]
propagate = "no"
[logging.loggers.emergency]
level = "DEBUG"
handlers = ["errorHandler"]
propagate = "no"
